ai-housekeeping-advisor/backend/image-processor/deploy/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.1.0
google-cloud-vision==3.10.1
pydantic==2.11.1
